requirements.txt
absl-py==0.14.1
addict==2.4.0
anyio==3.3.4
argcomplete==1.12.3
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
ConfigArgParse==1.5.3
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
dpcpp-cpp-rt==2022.1.0
entrypoints==0.3
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.4.0
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.1
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyter-packaging==0.11.0
jupyter-server==1.11.1
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lpips==0.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mkl==2022.1.0
mkl-fft==1.3.0
mkl-service==2.4.0
nbclassic==0.3.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.6
oauthlib==3.1.1
olefile==0.46
open3d==0.13.0
opencv-contrib-python==4.5.3.56
OpenEXR==1.3.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.18.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
tbb==2021.6.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.10.12
tomlkit==0.7.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
trimesh==3.13.5
typing_extensions==4.3.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
zipp==3.6.0
